Add files via upload
An example to understand why TBATS is good modeling only seasonal models and not random time series.
Experiment on SBI NSE Index . It is not seasonal. Hence when TBATS was applied final output was in accordance with the last year(considering the seasonal function) but not in accordance with the current values.